requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
python-dotenv
mcp
pytest
openai
google-generativeai
